NetworkAnalyzer: rework entire flowgraph algorithm
Rework the entire flowgraph that was previously used to compute the single bin DFT using the goertzel algorithm. Also make sure the flowgraph can retrieve DC Voltage information and also cut it if enabled.
Allow the user to force the Gain mode of the response channel or let it be set automatically by the NetworkAnalyzer.

Signed-off-by: Daniel Guramulta <daniel.guramulta@analog.com>